requirements.txt
lpips
brisque
torch
torchvision
numpy
Pillow
opencv-python
scipy
scikit-image
pandas
piq
nltk
PyWavelets